refactor: increase fuzzing space

+++ contracts/test/proofs/DeltaProof.t.sol
@@ -303,7 +303,7 @@ contract DeltaProofTest is Test {
         DeltaFuzzing.verify({proof: proof, instance: deltaAcc, verifyingKey: verifyingKey});
     }
 
-    function testFuzz_add_reverts_when_adding_a_non_curve_from_the_right(uint8 k, Delta.Point memory rhs) public {
+    function testFuzz_add_reverts_when_adding_a_non_curve_from_the_right(uint32 k, Delta.Point memory rhs) public {
         // Ensure that `rhs` is not on the curve.
         vm.assume(!EllipticCurve.isOnCurve({_x: rhs.x, _y: rhs.y, _aa: Delta._AA, _bb: Delta._BB, _pp: Delta._PP}));
 
@@ -319,7 +319,7 @@ contract DeltaProofTest is Test {
         lhs.add(rhs);
     }
 
-    function testFuzz_add_reverts_when_adding_zero_from_the_right(uint8 k) public {
+    function testFuzz_add_reverts_when_adding_zero_from_the_right(uint32 k) public {
         // Generate a random point on the curve.
         Delta.Point memory lhs = _mul({p: Delta.Point({x: Delta._GX, y: Delta._GY}), k: k});
         assertTrue(
@@ -334,7 +334,7 @@ contract DeltaProofTest is Test {
         lhs.add(zero);
     }
 
-    function testFuzz_add_adding_zero_from_the_left_produces_a_curve_point(uint8 k) public pure {
+    function testFuzz_add_adding_zero_from_the_left_produces_a_curve_point(uint32 k) public pure {
         Delta.Point memory zero = Delta.zero();
 
         // Generate a random point on the curve.
@@ -352,7 +352,7 @@ contract DeltaProofTest is Test {
         );
     }
 
-    function testFuzz_add_adding_zero_from_the_left_is_the_identity_operation(uint8 k) public pure {
+    function testFuzz_add_adding_zero_from_the_left_is_the_identity_operation(uint32 k) public pure {
         Delta.Point memory zero = Delta.zero();
 
         // Generate a random point on the curve.
@@ -368,7 +368,7 @@ contract DeltaProofTest is Test {
         assertEq(sum.y, rhs.y);
     }
 
-    function testFuzz_add_adding_two_curve_points_produces_a_curve_point(uint8 k1, uint8 k2) public pure {
+    function testFuzz_add_adding_two_curve_points_produces_a_curve_point(uint32 k1, uint32 k2) public pure {
         // Generate two random points on the curve.
         Delta.Point memory lhs = _mul({p: Delta.Point({x: Delta._GX, y: Delta._GY}), k: k1});
         assertTrue(
